+++ app/generation/engine.js
@@ -1,4 +1,4 @@
-const Generation = require('./generation'); 
+const Generation = require('.'); 
 
 class GenerationEngine {
     constructor() {
