+++ app/index.js
@@ -1,9 +1,8 @@
 const express = require('express');
-const GenerationEngine = require('./engine');
+const GenerationEngine = require('./generation/engine');
 
 const app = express();
 const engine = new GenerationEngine();
-const port = 3000;
 
 engine.start();
 
@@ -11,4 +10,4 @@ app.get('/dragon/new', (req, res) => {
     res.json({ dragon: engine.generation.newDragon() });
 });
 
-app.listen(port, () => console.log(`listening on port ${port}`));
+module.exports = app;
