+++ bin/server.js
@@ -1,0 +1,7 @@
+const app = require('../app/index');
+
+const port = 3000;
+
+app.listen(port, () => console.log(`listening on port ${port}`));
+
+
